requirements.txt
matplotlib
torchgeo==0.3.1
opencv-python==4.7.0.68
albumentations==1.3.0
torchsummary==1.5.1
ttach==0.0.3
earthengine-api
richdem==0.3.4
tqdm
opencv-python==4.7.*
scikit-learn==1.2.1
rasterio==1.3.4